Remove Address Extensions out of Core
