+++ src/Stratis.Bitcoin.Features.SmartContracts.Tests/AuctionTests.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using Stratis.Bitcoin.Features.SmartContracts.Networks;
 using Stratis.SmartContracts;
 using Stratis.SmartContracts.Core;
+using Stratis.SmartContracts.Executor.Reflection;
 using Stratis.SmartContracts.Executor.Reflection.Serialization;
 using Xunit;
