+++ src/Stratis.Bitcoin.Features.SmartContracts.Tests/ContractLogHolderTests.cs
@@ -6,6 +6,7 @@ using Stratis.Bitcoin.Features.SmartContracts.Networks;
 using Stratis.SmartContracts;
 using Stratis.SmartContracts.Core;
 using Stratis.SmartContracts.Core.Receipts;
+using Stratis.SmartContracts.Executor.Reflection;
 using Stratis.SmartContracts.Executor.Reflection.ContractLogging;
 using Stratis.SmartContracts.Executor.Reflection.Serialization;
 using Xunit;
