+++ src/Stratis.Bitcoin.Features.SmartContracts.Tests/ContractPrimitiveSerializerTests.cs
@@ -5,6 +5,7 @@ using Newtonsoft.Json;
 using Stratis.Bitcoin.Features.SmartContracts.Networks;
 using Stratis.SmartContracts;
 using Stratis.SmartContracts.Core;
+using Stratis.SmartContracts.Executor.Reflection;
 using Stratis.SmartContracts.Executor.Reflection.Exceptions;
 using Stratis.SmartContracts.Executor.Reflection.Serialization;
 using Xunit;
