+++ src/Stratis.Bitcoin.Features.SmartContracts.Tests/MethodParameterByteSerializerTests.cs
@@ -5,6 +5,7 @@ using System.Text;
 using NBitcoin;
 using Stratis.Bitcoin.Features.SmartContracts.Networks;
 using Stratis.SmartContracts.Core;
+using Stratis.SmartContracts.Executor.Reflection;
 using Stratis.SmartContracts.Executor.Reflection.Serialization;
 using Xunit;
