+++ src/Stratis.Bitcoin.Features.SmartContracts.Tests/MethodParameterStringSerializerTests.cs
@@ -6,6 +6,7 @@ using NBitcoin;
 using Stratis.Bitcoin.Features.SmartContracts.Networks;
 using Stratis.SmartContracts;
 using Stratis.SmartContracts.Core;
+using Stratis.SmartContracts.Executor.Reflection;
 using Stratis.SmartContracts.Executor.Reflection.Serialization;
 using Xunit;
