+++ src/Stratis.SmartContracts.Core/Extensions.cs
@@ -26,42 +26,7 @@ namespace Stratis.SmartContracts.Core
             return bytes;
         }
 
-        public static uint160 ToUint160(this Address address)
-        {
-            return new uint160(address.ToBytes());
-        }
-
-        public static uint160 ToUint160(this string base58Address, Network network)
-        {
-            return new uint160(new BitcoinPubKeyAddress(base58Address, network).Hash.ToBytes());
-        }
-
-        public static Address ToAddress(this uint160 address)
-        {
-            return Address.Create(address.ToBytes());
-        }
-
-        public static Address ToAddress(this string address, Network network)
-        {
-            return Address.Create(address.ToUint160(network).ToBytes());
-        }
-
-        public static Address ToAddress(this byte[] bytes)
-        {
-            if (bytes.Length != Address.Width)
-                throw new ArgumentOutOfRangeException(nameof(bytes), "Address must be 20 bytes wide");
-
-            return Address.Create(bytes);
-        }
-
-        public static Address HexToAddress(this string hexString)
-        {
-            // uint160 only parses a big-endian hex string
-            var result = HexStringToBytes(hexString);
-            return Address.Create(result);
-        }
-
-        private static byte[] HexStringToBytes(string val)
+        public static byte[] HexStringToBytes(string val)
         {
             if (val.StartsWith("0x", StringComparison.OrdinalIgnoreCase))
                 val = val.Substring(2);
@@ -73,11 +38,6 @@ namespace Stratis.SmartContracts.Core
                 ret[i / 2] = byte.Parse(hexChars, System.Globalization.NumberStyles.HexNumber);
             }
             return ret;
-        }
-
-        public static string ToBase58Address(this uint160 address, Network network)
-        {
-            return new BitcoinPubKeyAddress(new KeyId(address), network).ToString();
         }
 
         public static Money GetFee(this Transaction transaction, UnspentOutputSet inputs)
