+++ src/Stratis.SmartContracts.Token.Tests/StandardTokenTests.cs
@@ -2,6 +2,7 @@ using System;
 using Moq;
 using Stratis.Bitcoin.Features.SmartContracts.Networks;
 using Stratis.SmartContracts.Core;
+using Stratis.SmartContracts.Executor.Reflection;
 using Xunit;
 
 namespace Stratis.SmartContracts.Token.Tests
