+++ src/Stratis.SmartContracts/AssemblyInfo.cs
@@ -5,6 +5,7 @@
 [assembly: InternalsVisibleTo("Stratis.SmartContracts.Backend")]
 [assembly: InternalsVisibleTo("Stratis.SmartContracts.Demo")]
 [assembly: InternalsVisibleTo("Stratis.SmartContracts.Compiler")]
+[assembly: InternalsVisibleTo("Stratis.SmartContracts.Executor.Reflection")]
 [assembly: InternalsVisibleTo("Stratis.Bitcoin.Features.SmartContracts")]
 [assembly: InternalsVisibleTo("Stratis.Bitcoin.Features.SmartContractsApi")]
 [assembly: InternalsVisibleTo("Stratis.Bitcoin.Features.SmartContracts.Tests")]
